assume float scaler

+++ test/test_sat_cast.py
@@ -30,7 +30,7 @@ def test_cast(num_rows: int, num_cols: int, in_dtype: torch.dtype, fp8_dtype: to
     # but torch.rand is gaussian(0, 1) so it should be fine and we wont
     # exceed the range but we should do this right.
     a = torch.rand(num_rows, num_cols, dtype=in_dtype, device="cuda")
-    abs_max = torch.max(torch.abs(a))
+    abs_max = torch.max(torch.abs(a)).to(torch.float32)
 
     cast_pytorch = eager_scaled_quant(a, abs_max, fp8_dtype)
     cast_custom = saturated_cast(a, fp8_dtype, abs_max)
